% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{all_cluster_tissues}
\alias{all_cluster_tissues}
\title{Clusters tissues using all group means}
\usage{
all_cluster_tissues(gen, clusters, q = 6, G = 2)
}
\arguments{
\item{gen}{emmix-gene object}

\item{clusters}{mclust object}

\item{q}{number of factors if using mfa}

\item{G}{number of components if using mfa}
}
\value{
a clustering for each sample (columns) by each group(rows)
}
\description{
Clusters tissues using all group means
}
