% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_genes}
\alias{cluster_genes}
\title{Clusters genes}
\usage{
cluster_genes(gen, g = NULL)
}
\arguments{
\item{gen}{an emmix-gene object produced by select genes.}

\item{g}{number of gene clusters . If not specified will be selected automatically on the basis of BIC.}
}
\value{
mclust object
}
\description{
Clusters genes
}
\examples{

}
