% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_tissues}
\alias{cluster_tissues}
\title{Clusters tissues}
\usage{
cluster_tissues(genes, method = "t")
}
\arguments{
\item{genes}{matrix of genes}

\item{g}{number of clusters}
}
\value{
?
}
\description{
Clusters tissues
}
\examples{

}
