% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{heat_maps}
\alias{heat_maps}
\title{Heat maps}
\usage{
heat_maps(dat, clustering = NULL)
}
\arguments{
\item{dat}{matrix of gene expression data.}

\item{clustering}{a vector of sample classifications. MUst be same length as the number of columns in dat.}
}
\value{
A ggplot2 heat map.
}
\description{
Plot heat maps of gene expression data. Optionally sort the x-axis according to a predtermined clustering.
}
\examples{
data(alon_data)
example <- heat_maps(alon_data[1:100, ])

  
}
