% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_single_gene}
\alias{plot_single_gene}
\title{Plot single gene}
\usage{
plot_single_gene(dat, gene_id, random_starts = 50)
}
\arguments{
\item{dat}{full matrix of data}

\item{gene_id}{row number}

\item{random_starts}{number of random starts for the fit}
}
\value{
ggplot2 histogram with fit
}
\description{
Plot single gene
}
\examples{

}
