% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_genes}
\alias{select_genes}
\title{Selects Genes from data supplied.}
\usage{
select_genes(filename, random_starts = 4, max_it = 400, ll_thresh = 8,
  min_clust_size = 8)
}
\arguments{
\item{filename}{Name of file containing gene data.}
}
\value{
matrix of genes
}
\description{
Selects Genes from data supplied.
}
\examples{

}
