% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_genes_dat}
\alias{select_genes_dat}
\title{Selects Genes from data supplied.}
\usage{
select_genes_dat(filename, g = 1, k = 1, random_starts = 5,
  ll_thresh = 0.001, min_clust_size = 3)
}
\arguments{
\item{filename}{Name of file containing gene data.}

\item{g}{A number.}

\item{k}{A number.}
}
\value{
not sure yet
}
\description{
Selects Genes from data supplied.
}
