% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{top_genes_cluster_tissues}
\alias{top_genes_cluster_tissues}
\title{Clusters tissues}
\usage{
top_genes_cluster_tissues(gen, n_top = 100, method = "mfa", k = 2)
}
\arguments{
\item{gen}{emmix-gene object}

\item{n_top}{number of top genes (as ranked by liklihood) to be selected}

\item{method}{Method for seperating tissue classes. Can be either 't' for a univariate mixture of t-distributions on gene cluster means, or 'mfa' for a mixture of factor analysers.}

\item{k}{number of factors if using mfa}
}
\value{
a clustering for each sample (columns) by each group(rows)
}
\description{
Clusters tissues
}
\examples{

}
