% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{all_cluster_tissues}
\alias{all_cluster_tissues}
\title{Clusters tissues using all group means}
\usage{
all_cluster_tissues(gen, clusters, q = 6, G = 2)
}
\arguments{
\item{gen}{emmix-gene object}

\item{clusters}{mclust object}

\item{q}{number of factors if using mfa}

\item{G}{number of components if using mfa}
}
\value{
a clustering for each sample (columns) by each group(rows)
}
\description{
Clusters tissues using all group means
}
\examples{
data(alon_data)
#only run on first 100 genes for speed
alon_sel <- select_genes(alon_data[1:100, ]) 
alon_clust<- cluster_genes(alon_sel , 2)
alon_tissue_all<-all_cluster_tissues(alon_sel, alon_clust, q=1, G=2)
}
