% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_genes}
\alias{cluster_genes}
\title{Clusters genes}
\usage{
cluster_genes(genes, g = NULL)
}
\arguments{
\item{genes}{matrix of genes}

\item{g}{number of clusters}
}
\value{
mclust object
}
\description{
Clusters genes
}
\examples{

}
