% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_genes}
\alias{cluster_genes}
\title{Clusters genes using mixtures of normal distributions}
\usage{
cluster_genes(gen, g = NULL)
}
\arguments{
\item{gen}{an emmix-gene object produced by select_genes().}

\item{g}{The desired number of gene clusters. If not specified will be 
selected automatically on the basis of BIC.}
}
\value{
An array containing the clustering.
}
\description{
Sorts genes into clusters using mixtures of normal distributions with
covariance matrices restricted to be multiples of the indentity matrix.
}
\examples{
data(alon_data)
#only run on first 100 genes for speed
example <- select_genes(alon_data[1:100, ]) 
example2<- cluster_genes(example, 2)
}
