% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_tissues}
\alias{cluster_tissues}
\title{Clusters tissues}
\usage{
cluster_tissues(gen, clusters, method = "t", k = 2)
}
\arguments{
\item{gen}{emmix-gene object}

\item{clusters}{mclust object}

\item{method}{Method for seperating tissue classes. Can be either 't' for a univariate mixture of t-distributions on gene cluster means, or 'mfa' for a mixture of factor analysers.}

\item{k}{number of factors if using mfa}
}
\value{
a clustering for each sample (columns) by each group(rows)
}
\description{
Clusters tissues
}
\examples{

}
