% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{heat_maps}
\alias{heat_maps}
\title{Heat maps}
\usage{
heat_maps(dat, clustering = NULL, y_lab = NULL)
}
\arguments{
\item{dat}{matrix of gene expression data.}

\item{clustering}{a vector of sample classifications.
MUst be same length as the number of columns in dat.}

\item{y_lab}{optional label for y-axis.}
}
\value{
A ggplot2 heat map.
}
\description{
Plot heat maps of gene expression data. Optionally sort the x-axis 
according to a predtermined clustering.
}
\examples{
library(Biobase)
library(vsn)
set.seed(123)
data(colonCA, package = 'colonCA')
alon_data<-(Biobase::exprs(vsn::vsn2(colonCA))) 
example <- heat_maps(alon_data[seq_len(100), ])


}
