% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{heat_maps}
\alias{heat_maps}
\title{Heat maps}
\usage{
heat_maps(clust_genes, clustering = NULL)
}
\arguments{
\item{clust_genes}{matrix of genes}
}
\value{
ggplot2 heat map
}
\description{
Heat maps
}
\examples{

}
